Move model workload into each services . .
User, Post, and Group . .

+++ LanGeng.API/Dtos/FilterPostDto.cs
@@ -5,8 +5,8 @@ namespace LanGeng.API.Dtos;
 
 public record class FilterPostDto
 (
-    [Range(0, int.MaxValue)] int Page,
-    [Range(1, 64)] int? Limit = null,
+    [Range(0, short.MaxValue)] short Page,
+    [Range(1, 64)] byte? Limit = null,
     [StringLength(255)] string? Author = null,
     [StringLength(255)] string? Group = null,
     string? Keyword = null,
